+++ Sources/JiraAPIClient/JiraAPIClient.swift
@@ -6,30 +6,29 @@ import JiraAPI
 import os.log
 
 public class JiraAPIClient<AuthSession>: ObservableObject where AuthSession: AuthenticationSession {
-    public init(configuration: Configuration, logger: Logger? = nil) {
+    public init(configuration: Configuration, logger: Logger? = nil, authLogger: Logger? = nil) {
         self.config = configuration
         self.logger = logger
+        self.authLogger = authLogger ?? logger
     }
 
     internal let config: Configuration
-    lazy var auth: Auth = .init(doGetTokens: doGetTokens, doRefreshToken: doRefreshToken, logger: logger)
 
-    func doGetTokens() -> AnyPublisher<Auth.Tokens, Swift.Error> {
-        authorizationCode()
-            .flatMap(oauthTokens)
-            .eraseToAnyPublisher()
-    }
-
-    func doRefreshToken(refresh: Auth.Refresh) -> AnyPublisher<Auth.Tokens, Swift.Error> {
-        oauthTokensRefresh(for: refresh)
-            .eraseToAnyPublisher()
-    }
-
-    @Published var error: Swift.Error?
-
-    var signingIn: AnyCancellable?
+    lazy internal var auth: Auth = .init(doGetTokens: doGetTokens, doRefreshToken: doRefreshToken, logger: authLogger)
 
     var logger: Logger?
+    
+    var authLogger: Logger?
+
+    public var user: CurrentUser = .init()
+
+    public var resource: Resource?
+    
+    @Published public var error: Swift.Error?
+
+    internal var signingIn: AnyCancellable?
+    internal var fetchingCloudID: AnyCancellable?
+    internal var fetchingMyself: AnyCancellable?
 
     public func signIn() -> AnyPublisher<Never, Swift.Error> {
         auth.signIn()
@@ -37,18 +36,41 @@ public class JiraAPIClient<AuthSession>: ObservableObject where AuthSession: Aut
             .eraseToAnyPublisher()
     }
 
-    var fetchingCloudID: AnyCancellable?
+    @Published public var cloudID: JiraAPI.Auth.CloudID?
+
 
     public func getCloudID() {
         if let request = try? JiraAPI.Request.cloudResources() {
-
             fetchingCloudID = auth.fetch(request)
                 .map(\.data)
+                .decode(type: [JiraAPI.Models.CloudResourceResponse].self, decoder: JSONDecoder())
+                .map(\.first)
+                .map { $0! }
+                .map(Resource.init)
+                .receive(on: RunLoop.main)
                 .sink(receiveCompletion: { completion in
                     print("COMPLETED: \(completion)")
-                }, receiveValue: { data in
-                    let body = String(data: data, encoding: .utf8) ?? "nil"
-                    print("BODY: \(body)")
+                }, receiveValue: { [unowned self] resource in
+                    cloudID = resource.id
+                    user.update(with: [resource])
+                    print("RESOURCE NAME: \(resource.name)")
+                })
+        } else {
+            print("big error")
+        }
+    }
+
+    public func getMyself() {
+        if let cloudID = cloudID, let request = try? JiraAPI.Request.myself(cloudID: cloudID) {
+            fetchingCloudID = auth.fetch(request)
+                .map(\.data)
+                .decode(type: JiraAPI.Models.UserResponse.self, decoder: JSONDecoder())
+                .map(User.init)
+                .receive(on: RunLoop.main)
+                .sink(receiveCompletion: { completion in
+                    print("COMPLETED: \(completion)")
+                }, receiveValue: { [unowned self] in
+                    user.update(with: $0)
                 })
         } else {
             print("big error")