+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- separate logger for auth
+- current user object
 ### Changed
 - using logging data task from newer version of combine extras
 ### Deprecated